Add PGXN support and refactor template files
- Introduce pgxn module for PGXN package generation
- Remove unused license template and update README template
- Refactor package.rs to utilize new pgxn module for PGXN package creation
