+++ cargo-pgrx/src/templates/readme_md
@@ -1,86 +1,3 @@
 # {name}
 
 A PostgreSQL extension built with [pgrx](https://github.com/pgcentralfoundation/pgrx).
-
-## Overview
-
-Brief description of what this extension does.
-
-## Features
-
-- Feature 1
-- Feature 2
-- Feature 3
-
-## Building
-
-### Requirements
-
-- Rust (latest stable)
-- PostgreSQL development headers
-- pgrx development environment initialized via `cargo pgrx init`
-
-### Compilation
-
-```bash
-cd {name}
-cargo pgrx build
-```
-
-## Installation
-
-### Quick Install
-
-```bash
-cargo pgrx install
-```
-
-This will compile the extension and install it into a pgrx-managed PostgreSQL instance.
-
-### Production Install
-
-For production deployments:
-
-```bash
-cargo pgrx package
-```
-
-Then follow the standard PostgreSQL extension installation procedures.
-
-## Usage
-
-```sql
-CREATE EXTENSION {name};
-```
-
-## Testing
-
-Run the test suite:
-
-```bash
-cargo pgrx test
-```
-
-Test against specific PostgreSQL versions:
-
-```bash
-cargo pgrx test --pgXX
-```
-
-## Documentation
-
-For more information about:
-- Building PostgreSQL extensions with Rust: See [pgrx documentation](https://docs.rs/pgrx/)
-- PostgreSQL development: See [PostgreSQL documentation](https://www.postgresql.org/docs/)
-
-## License
-
-This extension is licensed under the MIT License. See LICENSE file for details.
-
-## Contributing
-
-Contributions are welcome! Please feel free to submit a Pull Request.
-
-## Support
-
-For issues, questions, or suggestions, please open an issue on the [GitHub repository](https://github.com/yourusername/{name}).
